refactor: useWindowSize hook
